python-server/requirements.txt
flask==2.0.1
dlib==19.24.2
flask-cors==3.0.10
flask-socketio==5.1.1
psycopg2==2.9.1
python-dotenv==0.19.1
python-dateutil==2.8.2
gTTS==2.2.3
SpeechRecognition==3.10.0
pydub==0.25.1
cryptography==3.4.8
discord.py==1.7.3
boto3==1.17.82
face-recognition==1.3.0
opencv-python==4.7.0.72
pyttsx3==2.90
PyAudio==0.2.13
python-espeak==0.6.3
PyJWT==2.7.0
imageai==3.0.3
soundfile==0.12.1
tensorflow==2.13.0
torch==2.0.1
